NO-TICKET: rename models
