+++ core/application/models/cognito/CognitoModel.ts
