+++ core/application/models/dbModels/AcceptDonationModel.ts
