+++ core/application/models/logger/Logger.ts
