+++ core/application/models/sns/SNSModel.ts
