+++ core/application/models/sqs/SQSModel.ts
