+++ core/application/tests/mocks/mockUserData.ts
@@ -1,5 +1,5 @@
 import { UserDetailsDTO, UserDTO } from "../../../../commons/dto/UserDTO";
-import { UserFields } from "../../models2/dbModels/UserModel";
+import { UserFields } from "../../models/dbModels/UserModel";
 
 export const mockUserWithStringId: UserDTO = {
   id: "12345",
