+++ core/application/tests/technicalImpl/dbModels/UserModel.test.ts
@@ -1,4 +1,4 @@
-import UserModel from "../../../models2/dbModels/UserModel";
+import UserModel from "../../../models/dbModels/UserModel";
 import {
   expectedUser,
   mockUserDetailsWithStringId,
