+++ core/services/aws/commons/cognito/CognitoOperations.ts
@@ -2,7 +2,7 @@ import {
   CognitoIdentityProviderClient,
   AdminUpdateUserAttributesCommand,
 } from "@aws-sdk/client-cognito-identity-provider";
-import { UpdateCognitoAttributes } from "../../../../application/models2/cognito/CognitoModel";
+import { UpdateCognitoAttributes } from "../../../../application/models/cognito/CognitoModel";
 
 const cognitoClient = new CognitoIdentityProviderClient({
   region: process.env.AWS_REGION,
