+++ core/services/aws/tests/commons/ddb/DynamoDbTableOperations.test.ts
@@ -8,14 +8,14 @@ import {
 } from "@aws-sdk/lib-dynamodb";
 import { mockClient } from "aws-sdk-client-mock";
 import { UserDetailsDTO } from "../../../../../../commons/dto/UserDTO";
-import UserModel from "../../../../../application/models2/dbModels/UserModel";
+import UserModel from "../../../../../application/models/dbModels/UserModel";
 import DatabaseError from "../../../../../../commons/libs/errors/DatabaseError";
 import { GENERIC_CODES } from "../../../../../../commons/libs/constants/GenericCodes";
 import {
   mockUserDetailsWithStringId,
   expectedUser,
 } from "../../../../../application/tests/mocks/mockUserData";
-import { QueryConditionOperator } from "../../../../../application/models2/policies/repositories/QueryTypes";
+import { QueryConditionOperator } from "../../../../../application/models/policies/repositories/QueryTypes";
 
 describe("DynamoDbTableOperations Tests", () => {
   const ddbMock = mockClient(DynamoDBDocumentClient);
